+++ CHANGES.rst
@@ -115,6 +115,17 @@
 - Make ``SimpleVocabulary`` and ``SimpleTerm`` have value-based
   equality and hashing methods.
 
+- All fields of the schema of an ``Object`` field are bound to the
+  top-level value being validated before attempting validation of
+  their particular attribute. Previously only ``IChoice`` fields were
+  bound. See `issue 17
+  <https://github.com/zopefoundation/zope.schema/issues/17>`_.
+
+- Share the internal logic of ``Object`` field validation and
+  ``zope.schema.getValidationErrors``. See `issue 57
+  <https://github.com/zopefoundation/zope.schema/issues/57>`_.
+
+
 4.5.0 (2017-07-10)
 ==================